requirements.txt
torch==2.5.1
torchvision==0.20.1
streamlit
pillow
scikit-learn
matplotlib
seaborn
segmentation-models-pytorch
timm
ttach